fastapi-notes/26-gyansetu/requirements.txt
fastapi[standard]>=0.110.0
uvicorn>=0.27.0
google-generativeai>=0.4.0
chromadb>=0.4.22
sqlmodel>=0.0.14
pydantic-settings>=2.1.0
python-multipart>=0.0.6
